Adding an if-else statement to cover missing images
Introducing an if-else statement to cover the eventuality where an image is not present on disk. A check is made to see if the file exists on disk; if found, then a directory is created and the copied, else a warning message is logged, specifying the ImageID of the missing file and where it was expected to be found.